добавил OfferParser
