+++ src/shared/libs/offer-parser/index.ts
@@ -1,0 +1,2 @@
+export { OfferParser } from './offer-parser.interface.js';
+export { TSVOfferParser } from './tsv-offer-parser.js';
