+++ src/shared/libs/offer-parser/offer-parser.interface.ts
@@ -1,0 +1,5 @@
+import { Offer, OfferId } from '#shared/types/offer.type.js';
+
+export interface OfferParser {
+  parse(line: string, id: OfferId): Offer;
+}
